Move constants and reprecation utilities into private utils module
