+++ orix/_utils/__init__.py
@@ -17,20 +17,4 @@
 # along with orix. If not, see <http://www.gnu.org/licenses/>.
 #
 
-import pytest
-
-from orix._utils.constants import installed
-
-
-class TestConstants:
-    @pytest.mark.skipif(
-        installed["numpy-quaternion"], reason="numpy-quaternion is installed"
-    )
-    def test_numpy_quaternion_not_installed(self):
-        assert not installed["numpy-quaternion"]
-
-    @pytest.mark.skipif(
-        not installed["numpy-quaternion"], reason="numpy-quaternion is installed"
-    )
-    def test_numpy_quaternion_installed(self):
-        assert installed["numpy-quaternion"]
+"""Private utilities used across modules."""
