+++ orix/_utils/constants.py
